requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
djangorestframework==3.12.4
importlib-metadata==4.8.1
kombu==5.1.0
prompt-toolkit==3.0.20
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
